Refactor: fix-up namespaces
