+++ NumberTranslator.Tests/Translators/EnglishNumberTranslatorTests.cs
@@ -1,6 +1,6 @@
 using System.Globalization;
 using FluentAssertions;
-using NumberTranslator.Translators;
+using NumberTranslator.Translators.Languages;
 
 namespace NumberTranslator.Tests.Translators;
